requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.14.0
Flask==3.0.3
fsspec==2024.5.0
ftfy==6.2.0
huggingface-hub==0.23.1
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
open-clip-torch==2.24.0
packaging==24.0
pillow==10.3.0
protobuf==5.27.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
sentencepiece==0.2.0
sympy==1.12
tbb==2021.12.0
timm==1.0.3
tokenizers==0.15.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.38.2
typing_extensions==4.12.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
